+++ src/gui/PrivilegesDialog.cpp
@@ -505,7 +505,7 @@ void PrivilegesDialog::updateControls()
 wxString PrivilegesDialog::getSqlStatements()
 {
     wxString stmt;
-    for (int i = 0; i < listbox_statements->GetCount(); i++)
+    for (int i = 0; i < (int)listbox_statements->GetCount(); i++)
         stmt << listbox_statements->GetString(i) << wxT(";\n");
     return stmt;
 }
